requirements.txt
PyPDF2
sentence-transformers
faiss-cpu
numpy
transformers
torch
